extract makeTweetstorm
